# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/registrations/clients": {
      "post": {
        "operationId": "RegistrationController_registerClient",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneChallengeDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or OTP cooldown; retryAt is supplied for cooldown.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Submit a Client registration and send a phone code",
        "tags": [
          "registration"
        ]
      }
    },
    "/v1/registrations/service-providers": {
      "post": {
        "description": "Requires the existing business/contact fields including email. Unverified pending retries resume the stored profile without replacing it. Approval is decided only after phone verification by server policy.",
        "operationId": "RegistrationController_registerServiceProvider",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceProviderRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneChallengeDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or OTP cooldown; retryAt is supplied for cooldown.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Submit a password-free Service Provider registration and send a phone code",
        "tags": [
          "registration"
        ]
      }
    },
    "/v1/registrations/verify-phone": {
      "post": {
        "operationId": "RegistrationController_verifyPhone",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPhoneCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationVerificationDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Verify a registration phone number",
        "tags": [
          "registration"
        ]
      }
    },
    "/v1/registrations/resend-phone-code": {
      "post": {
        "operationId": "RegistrationController_resendPhoneCode",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhoneChallengeDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or OTP cooldown; retryAt is supplied for cooldown.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "summary": "Replace and resend a registration phone code",
        "tags": [
          "registration"
        ]
      }
    },
    "/v1/provider-registrations/pending": {
      "get": {
        "operationId": "ProviderRegistrationController_pending",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PendingProviderRegistrationDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List phone-verified Provider registrations awaiting review",
        "tags": [
          "provider registration review"
        ]
      }
    },
    "/v1/provider-registrations": {
      "get": {
        "operationId": "ProviderRegistrationController_providers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProviderAdministrationDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List Providers by administration status",
        "tags": [
          "provider registration review"
        ]
      }
    },
    "/v1/provider-registrations/{providerUserId}": {
      "get": {
        "operationId": "ProviderRegistrationController_detail",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderAdministrationDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "provider registration review"
        ]
      }
    },
    "/v1/provider-registrations/{providerUserId}/status": {
      "post": {
        "operationId": "ProviderRegistrationController_changeStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderAdministrationDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "provider registration review"
        ]
      }
    },
    "/v1/provider-registrations/{providerUserId}/decision": {
      "post": {
        "operationId": "ProviderRegistrationController_decide",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderRegistrationStatusDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve or reject a pending Provider registration",
        "tags": [
          "provider registration review"
        ]
      }
    },
    "/v1/provider-registrations/me/status": {
      "get": {
        "operationId": "ProviderRegistrationController_status",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderRegistrationStatusDto"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read the current Provider registration status",
        "tags": [
          "provider registration review"
        ]
      }
    },
...
  },
  "info": {
    "title": "Weyonje API",
    "description": "Native authentication, server-managed sessions, and authoritative eligibility for Weyonje clients.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "PhoneChallengeDto": {
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          },
          "maskedPhone": {
            "type": "string",
            "example": "+256 •••••• 123"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "resendAvailableAt": {
            "type": "string",
            "format": "date-time"
          },
          "deliveryStatus": {
            "type": "string",
            "enum": [
              "SENT",
              "FAILED"
            ]
          }
        },
        "required": [
          "challengeId",
          "maskedPhone",
          "expiresAt",
          "resendAvailableAt",
          "deliveryStatus"
        ]
      },
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "retryAt": {
            "type": "string",
            "format": "date-time"
          },
          "limitCategory": {
            "type": "string",
            "enum": [
              "OTP_HOURLY",
              "OTP_COOLDOWN",
              "CLIENT_REQUEST",
              "PROVIDER_REQUEST"
            ]
          },
          "code": {
            "type": "string",
            "example": "AUTH_ACCESS_DENIED"
          },
          "message": {
            "type": "string",
            "example": "This account cannot access Weyonje mobile services."
          },
          "requestId": {
            "type": "string",
            "example": "2e96d4c8-5f83-4de0-957e-a7f0cff86690"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "VerifyPhoneCodeDto": {
        "type": "object",
        "properties": {
          "challengeId": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string",
            "pattern": "^[0-9]{6}$",
            "writeOnly": true
          }
        },
        "required": [
          "challengeId",
          "code"
        ]
      },
...
      "ClientRegistrationDto": {
        "type": "object",
        "properties": {
          "clientType": {
            "type": "string",
            "enum": [
              "INDIVIDUAL",
              "ORGANIZATION"
            ]
          },
          "firstName": {
            "type": "string",
            "maxLength": 100
          },
          "lastName": {
            "type": "string",
            "maxLength": 100
          },
          "organizationName": {
            "type": "string",
            "maxLength": 200
          },
          "phoneNumber": {
            "type": "string",
            "example": "+256 700 000000",
            "maxLength": 40
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "contactPersonName": {
            "type": "string",
            "maxLength": 200
          },
          "contactPersonPhone": {
            "type": "string",
            "maxLength": 40
          }
        },
        "required": [
          "clientType",
          "phoneNumber"
        ]
      },
      "ServiceProviderRegistrationDto": {
        "type": "object",
        "properties": {
          "essLicenseNumber": {
            "type": "string",
            "maxLength": 100
          },
          "companyName": {
            "type": "string",
            "maxLength": 200
          },
          "phoneNumber": {
            "type": "string",
            "maxLength": 40
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "workAddress": {
            "type": "string",
            "maxLength": 500
          },
          "providerType": {
            "type": "string",
            "enum": [
              "GULPER",
              "EMPTIER"
            ]
          },
          "contactPersonName": {
            "type": "string",
            "maxLength": 200
          },
          "contactPersonPhone": {
            "type": "string",
            "maxLength": 40
          }
        },
        "required": [
          "essLicenseNumber",
          "companyName",
          "phoneNumber",
          "email",
          "workAddress",
          "providerType",
          "contactPersonName",
          "contactPersonPhone"
        ]
      },
      "RegistrationVerificationDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "refreshTokenExpiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "accessTokenExpiresAt",
          "refreshTokenExpiresAt"
        ]
      },
      "PendingProviderRegistrationDto": {
        "type": "object",
        "properties": {
          "providerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "essLicenseNumber": {
            "type": "string"
          },
          "companyName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "workAddress": {
            "type": "string"
          },
          "providerType": {
            "type": "string",
            "enum": [
              "GULPER",
              "EMPTIER"
            ]
          },
          "contactPersonName": {
            "type": "string"
          },
          "contactPersonPhone": {
            "type": "string"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "providerUserId",
          "essLicenseNumber",
          "companyName",
          "phoneNumber",
          "email",
          "workAddress",
          "providerType",
          "contactPersonName",
          "contactPersonPhone",
          "submittedAt"
        ]
      },
      "ProviderStatusHistoryDto": {
        "type": "object",
        "properties": {
          "provenance": {
            "type": "string",
            "enum": [
              "KCCA_MANUAL",
              "SYSTEM_REGISTRATION_POLICY"
            ]
          },
          "fromStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED",
              "INACTIVE",
              "DISABLED"
            ]
          },
          "toStatus": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED",
              "INACTIVE",
              "DISABLED"
            ]
          },
          "reason": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "provenance",
          "fromStatus",
          "toStatus",
          "createdAt"
        ]
      },
      "ProviderAdministrationDto": {
        "type": "object",
        "properties": {
          "providerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "essLicenseNumber": {
            "type": "string"
          },
          "companyName": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "workAddress": {
            "type": "string"
          },
          "providerType": {
            "type": "string",
            "enum": [
              "GULPER",
              "EMPTIER"
            ]
          },
          "contactPersonName": {
            "type": "string"
          },
          "contactPersonPhone": {
            "type": "string"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED",
              "INACTIVE",
              "DISABLED"
            ]
          },
          "active": {
            "type": "boolean"
          },
          "providerNumber": {
            "type": "string"
          },
          "rejectionReason": {
            "type": "string"
          },
          "statusHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderStatusHistoryDto"
            }
          }
        },
        "required": [
          "providerUserId",
          "essLicenseNumber",
          "companyName",
          "phoneNumber",
          "email",
          "workAddress",
          "providerType",
          "contactPersonName",
          "contactPersonPhone",
          "submittedAt",
          "status",
          "active",
          "statusHistory"
        ]
      },
      "ProviderRegistrationStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "APPROVED",
              "REJECTED",
              "INACTIVE",
              "DISABLED"
            ]
          },
          "providerNumber": {
            "type": "string"
          },
          "rejectionReason": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}